app/career-counselling/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.8
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.7.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
Faker==37.4.0
fastapi==0.115.12
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-MySQLdb==2.0.0
fonttools==4.48.1
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
googleapis-common-protos==1.70.0
gunicorn==21.2.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
isort==6.0.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jose==1.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.3.1
MarkupSafe==2.1.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
motor==3.7.0
moviepy==1.0.3
multidict==6.4.3
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.3
passlib==1.7.4
pillow==10.4.0
platformdirs==4.2.0
pluggy==1.5.0
prettytable==3.8.0
proglog==0.1.10
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycurl==7.45.6
pydantic==2.11.3
pydantic-settings==2.9.0
pydantic_core==2.33.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.17.2
PyJWT==2.10.1
pylint==3.3.4
pymongo==4.13.2
PyMySQL==1.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
twilio==9.10.2
python-multipart==0.0.20
pytz==2025.1
pyzmq==25.1.2
razorpay==1.4.2
requests==2.32.3
RNBIP==1.6.1
rsa==4.9.1
scipy==1.15.2
selenium==4.27.1
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.46.2
tabulate==0.9.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.27.0
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
wikipedia==1.4.0
wptools==0.4.17
wsproto==1.2.0
yarl==1.19.0
python-socketio
fastapi-mail
